Refactor the code to seperated files JC-623
