+++ apps/vanilla/src/pages/blackjack/script/Player.ts
@@ -1,0 +1,23 @@
+import { Subscriber, Publisher } from './pattern';
+import { PlayerTurnResult } from './ResultDisplayer';
+
+export class Player implements Subscriber<PlayerTurnResult> {
+	private diceResult: number[] = [];
+	public result: Publisher<number[]> = new Publisher<number[]>();
+	public winStatus: Publisher<boolean> = new Publisher<boolean>();
+
+	public constructor(private playerIndex: number) {}
+
+	private calculateTotalDice = (): number => {
+		return this.diceResult.reduce((result, current) => result + current, 0);
+	};
+
+	update(turnResult: PlayerTurnResult) {
+		if (turnResult.playerIndex === this.playerIndex) this.diceResult.push(turnResult.diceResult);
+		this.result.notify([turnResult.diceResult]);
+		const totalDiceResult = this.calculateTotalDice();
+		if (totalDiceResult >= 21) {
+			this.winStatus.notify(true);
+		}
+	}
+}
